get results now requires a CL option to be set

+++ save_page_as.sh
@@ -19,6 +19,7 @@ destination="."
 browser="google-chrome"
 suffix=""
 url=""
+get_pages_number=false # CUSTOM VARIABLE (USED AT END)
 
 function print_usage() {
     printf "\n%s: Open the given url in a browser tab/window, perform 'Save As' operation and close the tab/window.\n\n" "${scriptname}" >&2
@@ -30,6 +31,7 @@ function print_usage() {
     printf "  -b, --browser          Browser executable to be used (must be one of 'google-chrome', 'chromium-browser' or 'firefox'). Default = '%s'.\n" "${browser}" >&2
     printf "  --load-wait-time       Number of seconds to wait for the page to be loaded (i.e., seconds to sleep before Ctrl+S is 'pressed'). Default = %s\n" "${load_wait_time}" >&2
     printf "  --save-wait-time       Number of seconds to wait for the page to be saved (i.e., seconds to sleep before Ctrl+F4 is 'pressed'). Default = %s\n" "${save_wait_time}" >&2
+    printf "  --get-pages-number    Scrape the number of pages of listings from the clue url just retrieved and save it to a text file. Default = %s\n" "${get_pages_number}" >&2
     printf "  -h, --help             Display this help message and exit.\n" >&2
 }
 
@@ -65,6 +67,11 @@ do
         -h | --help)
             print_usage
             exit 0
+            ;;
+        --get-pages-number) # CUSTOM OPTION
+            shift;
+            get_pages_number=true
+            shift
             ;;
         -*)
             printf "SAVE-AS: ERROR: Unknown option: %s\n" "${1}">&2
@@ -226,37 +233,44 @@ printf "SAVE-AS: INFO: Saving web page ...\n" >&2
 sleep ${save_wait_time}
 printf "SAVE-AS: INFO: Saved.\n">&2
 
-# Get number of results
-printf "SAVE-AS: INFO: Getting number of results...\n" >&2
-sleep 1s #
-xdotool key --clearmodifiers F12
-sleep 5s #
-xdotool key --clearmodifiers Tab
-sleep 1s #
-xdotool type --delay 50 --clearmodifiers "cell search-message first-cell"
-sleep 3s #
-xdotool key --clearmodifiers Return
-sleep 1s #
-xdotool key  --clearmodifiers Tab
-sleep 1s #
-xdotool key --clearmodifiers Tab
-sleep 1s #
-xdotool key --clearmodifiers Tab
-sleep 1s #
-xdotool key --clearmodifiers Right
-sleep 1s #
-xdotool key --clearmodifiers Down
-sleep 1s #
-xdotool key --clearmodifiers "ctrl+c"
+# Get number of pages of results with this clue
+if [ "$get_pages_number" = true ]
+then
+    printf "SAVE-AS: INFO: Getting number of results...\n" >&2
+    sleep 1s #
+    xdotool key --clearmodifiers F12
+    sleep 5s #
+    xdotool key --clearmodifiers Tab
+    sleep 1s #
+    xdotool type --delay 50 --clearmodifiers "cell search-message first-cell"
+    sleep 3s #
+    xdotool key --clearmodifiers Return
+    sleep 1s #
+    xdotool key  --clearmodifiers Tab
+    sleep 1s #
+    xdotool key --clearmodifiers Tab
+    sleep 1s #
+    xdotool key --clearmodifiers Tab
+    sleep 1s #
+    xdotool key --clearmodifiers Right
+    sleep 1s #
+    xdotool key --clearmodifiers Down
+    sleep 1s #
+    xdotool key --clearmodifiers "ctrl+c"
 
-# Process number
-resultsnumber=$(xclip -out -selection clipboard)
-pagesnumber=$(python -c "from math import ceil; print ceil($resultsnumber/25)")
+    # Process number
+    resultshtml=$(xclip -out -selection clipboard)
+    # Get integer substring
+    resultsnumber=$(echo $resultshtml| cut -d'>' -f 2)
+    resultsnumber=$(echo $resultsnumber| cut -d' ' -f 1)
+    # Do the math
+    pagesnumber=$(python -c "from math import ceil; print int(ceil($resultsnumber/35))")
 
-# Paste to fresh file
-rm ~/Documents/Newscrape/Binaries/pagesnumber.txt
-pagesnumber >> ~/Documents/Newscrape/Binaries/pagesnumber.txt
-printf "SAVE-AS: INFO: Got number of results.\n">&2
+    # Paste to fresh file
+    rm ~/Documents/Newscrape/Binaries/pagesnumber.txt
+    echo $pagesnumber >> ~/Documents/Newscrape/Binaries/pagesnumber.txt
+    printf "SAVE-AS: INFO: Got number of results.\n">&2
+fi
 
 # Close the browser tab/window (Ctrl+w for KDE, Ctrl+F4 otherwise)
 if [[ "${is_kde}" -eq 1 ]]; then
